Clean up print_Response for team
