+++ cookbook_v2/teams/concepts/streaming/01_team_streaming.py
@@ -49,4 +49,4 @@ team = Team(
 )
 
 # Test streaming response
-team.print_response("What is the current stock price of NVDA?", stream=True)
+team.print_response("What is the current stock price of NVDA?", stream=True, stream_intermediate_steps=True)
