+++ libs/agno/agno/team/team.py
@@ -44,6 +44,7 @@ from agno.run.messages import RunMessages
 from agno.run.response import RunEvent, RunOutput, RunOutputEvent
 from agno.run.team import TeamRunEvent, TeamRunOutput, TeamRunOutputEvent, ToolCallCompletedEvent
 from agno.session import SessionSummaryManager, TeamSession
+from agno.team.print import aprint_response, aprint_response_stream, print_response, print_response_stream
 from agno.tools.function import Function
 from agno.tools.toolkit import Toolkit
 from agno.utils.events import (
@@ -2286,7 +2287,8 @@ class Team:
             self.markdown = markdown
 
         if stream:
-            self._print_response_stream(
+            print_response_stream(
+                team=self,
                 input=input,
                 console=console,
                 show_message=show_message,
@@ -2306,7 +2308,8 @@ class Team:
                 **kwargs,
             )
         else:
-            self._print_response(
+            print_response(
+                team=self,
                 input=input,
                 console=console,
                 show_message=show_message,
@@ -2324,811 +2327,6 @@ class Team:
                 knowledge_filters=knowledge_filters,
                 **kwargs,
             )
-
-    def _print_response(
-        self,
-        input: Optional[Union[List, Dict, str, Message, BaseModel, List[Message]]] = None,
-        console: Optional[Any] = None,
-        show_message: bool = True,
-        show_reasoning: bool = True,
-        show_full_reasoning: bool = False,
-        tags_to_include_in_markdown: Optional[Set[str]] = None,
-        session_id: Optional[str] = None,
-        session_state: Optional[Dict[str, Any]] = None,
-        user_id: Optional[str] = None,
-        audio: Optional[Sequence[Audio]] = None,
-        images: Optional[Sequence[Image]] = None,
-        videos: Optional[Sequence[Video]] = None,
-        files: Optional[Sequence[File]] = None,
-        markdown: bool = False,
-        knowledge_filters: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> None:
-        import textwrap
-
-        from rich.console import Group
-        from rich.json import JSON
-        from rich.live import Live
-        from rich.markdown import Markdown
-        from rich.status import Status
-        from rich.text import Text
-
-        from agno.utils.response import format_tool_calls
-
-        if not tags_to_include_in_markdown:
-            tags_to_include_in_markdown = {"think", "thinking"}
-
-        with Live(console=console) as live_console:
-            status = Status("Thinking...", spinner="aesthetic", speed=0.4, refresh_per_second=10)
-            live_console.update(status)
-
-            response_timer = Timer()
-            response_timer.start()
-            # Panels to be rendered
-            panels = [status]
-            # First render the message panel if the message is not None
-            if input and show_message:
-                # Convert message to a panel
-                message_content = get_text_from_message(input)
-                message_panel = create_panel(
-                    content=Text(message_content, style="green"),
-                    title="Message",
-                    border_style="cyan",
-                )
-                panels.append(message_panel)
-                live_console.update(Group(*panels))
-
-            # Run the agent
-            run_response: TeamRunOutput = self.run(  # type: ignore
-                input=input,
-                images=images,
-                audio=audio,
-                videos=videos,
-                files=files,
-                stream=False,
-                session_id=session_id,
-                session_state=session_state,
-                user_id=user_id,
-                knowledge_filters=knowledge_filters,
-                **kwargs,
-            )
-            response_timer.stop()
-
-            team_markdown = False
-            member_markdown = {}
-            if markdown:
-                for member in self.members:
-                    if isinstance(member, Agent) and member.id is not None:
-                        member_markdown[member.id] = True
-                    if isinstance(member, Team) and member.id is not None:
-                        member_markdown[member.id] = True
-                team_markdown = True
-
-            if self.response_model is not None:
-                team_markdown = False
-
-            for member in self.members:
-                if member.response_model is not None and isinstance(member, Agent) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-                if member.response_model is not None and isinstance(member, Team) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-
-            # Handle reasoning
-            reasoning_steps = []
-            if (
-                isinstance(run_response, TeamRunOutput)
-                and run_response.metadata is not None
-                and run_response.metadata.reasoning_steps is not None
-            ):
-                reasoning_steps = run_response.metadata.reasoning_steps
-
-            if len(reasoning_steps) > 0 and show_reasoning:
-                # Create panels for reasoning steps
-                for i, step in enumerate(reasoning_steps, 1):
-                    reasoning_panel = self._build_reasoning_step_panel(i, step, show_full_reasoning)
-                    panels.append(reasoning_panel)
-                live_console.update(Group(*panels))
-
-            if isinstance(run_response, TeamRunOutput) and run_response.thinking is not None:
-                # Create panel for thinking
-                thinking_panel = create_panel(
-                    content=Text(run_response.thinking),
-                    title=f"Thinking ({response_timer.elapsed:.1f}s)",
-                    border_style="green",
-                )
-                panels.append(thinking_panel)
-                live_console.update(Group(*panels))
-
-            if isinstance(run_response, TeamRunOutput):
-                # Handle member responses
-                if self.show_members_responses:
-                    for member_response in run_response.member_responses:
-                        # Handle member reasoning
-                        reasoning_steps = []
-                        if (
-                            isinstance(member_response, RunOutput)
-                            and member_response.metadata is not None
-                            and member_response.metadata.reasoning_steps is not None
-                        ):
-                            reasoning_steps.extend(member_response.metadata.reasoning_steps)
-
-                        if len(reasoning_steps) > 0 and show_reasoning:
-                            # Create panels for reasoning steps
-                            for i, step in enumerate(reasoning_steps, 1):
-                                member_reasoning_panel = self._build_reasoning_step_panel(
-                                    i, step, show_full_reasoning, color="magenta"
-                                )
-                                panels.append(member_reasoning_panel)
-
-                        # Add tool calls panel for member if available
-                        if hasattr(member_response, "tools") and member_response.tools:
-                            member_name = None
-                            if isinstance(member_response, RunOutput) and member_response.id is not None:
-                                member_name = self._get_member_name(member_response.id)
-                            elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                                member_name = self._get_member_name(member_response.id)
-
-                            if member_name:
-                                formatted_calls = format_tool_calls(member_response.tools)
-                                if formatted_calls:
-                                    console_width = console.width if console else 80
-                                    panel_width = console_width + 30
-
-                                    lines = []
-                                    for call in formatted_calls:
-                                        wrapped_call = textwrap.fill(
-                                            f"• {call}", width=panel_width, subsequent_indent="  "
-                                        )
-                                        lines.append(wrapped_call)
-
-                                    tool_calls_text = "\n\n".join(lines)
-
-                                    member_tool_calls_panel = create_panel(
-                                        content=tool_calls_text,
-                                        title=f"{member_name} Tool Calls",
-                                        border_style="yellow",
-                                    )
-                                    panels.append(member_tool_calls_panel)
-                                    live_console.update(Group(*panels))
-
-                        show_markdown = False
-                        if member_markdown:
-                            if isinstance(member_response, RunOutput) and member_response.id is not None:
-                                show_markdown = member_markdown.get(member_response.id, False)
-                            elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                                show_markdown = member_markdown.get(member_response.id, False)
-
-                        member_response_content: Union[str, JSON, Markdown] = self._parse_response_content(
-                            member_response,
-                            tags_to_include_in_markdown,
-                            show_markdown=show_markdown,
-                        )
-
-                        # Create panel for member response
-                        if isinstance(member_response, RunOutput) and member_response.id is not None:
-                            member_response_panel = create_panel(
-                                content=member_response_content,
-                                title=f"{self._get_member_name(member_response.agent_id)} Response",
-                                border_style="magenta",
-                            )
-                        elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                            member_response_panel = create_panel(
-                                content=member_response_content,
-                                title=f"{self._get_member_name(member_response.team_id)} Response",
-                                border_style="magenta",
-                            )
-                        panels.append(member_response_panel)
-
-                        if member_response.citations is not None and member_response.citations.urls is not None:
-                            md_content = "\n".join(
-                                f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                                for i, citation in enumerate(member_response.citations.urls)
-                                if citation.url  # Only include citations with valid URLs
-                            )
-                            if md_content:  # Only create panel if there are citations
-                                citations_panel = create_panel(
-                                    content=Markdown(md_content),
-                                    title="Citations",
-                                    border_style="magenta",
-                                )
-                                panels.append(citations_panel)
-
-                    live_console.update(Group(*panels))
-
-                # Add team level tool calls panel if available
-                if run_response.tools:
-                    formatted_calls = format_tool_calls(run_response.tools)
-                    if formatted_calls:
-                        console_width = console.width if console else 80
-                        # Allow for panel borders and padding
-                        panel_width = console_width + 30
-
-                        lines = []
-                        for call in formatted_calls:
-                            wrapped_call = textwrap.fill(
-                                f"• {call}", width=panel_width, subsequent_indent="  "
-                            )  # Indent continuation lines
-                            lines.append(wrapped_call)
-
-                        # Join with blank lines between items
-                        tool_calls_text = "\n\n".join(lines)
-
-                        team_tool_calls_panel = create_panel(
-                            content=tool_calls_text,
-                            title="Team Tool Calls",
-                            border_style="yellow",
-                        )
-                        panels.append(team_tool_calls_panel)
-                        live_console.update(Group(*panels))
-
-                response_content_batch: Union[str, JSON, Markdown] = self._parse_response_content(
-                    run_response, tags_to_include_in_markdown, show_markdown=team_markdown
-                )
-
-                # Create panel for response
-                response_panel = create_panel(
-                    content=response_content_batch,
-                    title=f"Response ({response_timer.elapsed:.1f}s)",
-                    border_style="blue",
-                )
-                panels.append(response_panel)
-
-                # Add citations
-                if run_response.citations is not None and run_response.citations.urls is not None:
-                    md_content = "\n".join(
-                        f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                        for i, citation in enumerate(run_response.citations.urls)
-                        if citation.url  # Only include citations with valid URLs
-                    )
-                    if md_content:  # Only create panel if there are citations
-                        citations_panel = create_panel(
-                            content=Markdown(md_content),
-                            title="Citations",
-                            border_style="green",
-                        )
-                        panels.append(citations_panel)
-
-                if self.memory_manager is not None:
-                    if self.memory_manager.memories_updated:
-                        memory_panel = create_panel(
-                            content=Text("Memories updated"),
-                            title="Memories",
-                            border_style="green",
-                        )
-                        panels.append(memory_panel)
-
-                if self.session_summary_manager is not None and self.session_summary_manager.summaries_updated:
-                    summary_panel = create_panel(
-                        content=Text("Session summary updated"),
-                        title="Session Summary",
-                        border_style="green",
-                    )
-                    panels.append(summary_panel)
-                    self.session_summary_manager.summaries_updated = False
-
-            # Final update to remove the "Thinking..." status
-            panels = [p for p in panels if not isinstance(p, Status)]
-            live_console.update(Group(*panels))
-
-    def _print_response_stream(
-        self,
-        input: Optional[Union[List, Dict, str, Message, BaseModel, List[Message]]] = None,
-        console: Optional[Any] = None,
-        show_message: bool = True,
-        show_reasoning: bool = True,
-        show_full_reasoning: bool = False,
-        tags_to_include_in_markdown: Optional[Set[str]] = None,
-        session_id: Optional[str] = None,
-        session_state: Optional[Dict[str, Any]] = None,
-        user_id: Optional[str] = None,
-        audio: Optional[Sequence[Audio]] = None,
-        images: Optional[Sequence[Image]] = None,
-        videos: Optional[Sequence[Video]] = None,
-        files: Optional[Sequence[File]] = None,
-        messages: Optional[Sequence[Union[Dict, Message]]] = None,
-        markdown: bool = False,
-        stream_intermediate_steps: bool = False,  # type: ignore
-        knowledge_filters: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> None:
-        import textwrap
-
-        from rich.console import Group
-        from rich.json import JSON
-        from rich.live import Live
-        from rich.markdown import Markdown
-        from rich.status import Status
-        from rich.text import Text
-
-        from agno.utils.response import format_tool_calls
-
-        if not tags_to_include_in_markdown:
-            tags_to_include_in_markdown = {"think", "thinking"}
-
-        stream_intermediate_steps = True  # With streaming print response, we need to stream intermediate steps
-
-        _response_content: str = ""
-        _response_thinking: str = ""
-        reasoning_steps: List[ReasoningStep] = []
-
-        # Track tool calls by member and team
-        member_tool_calls = {}  # type: ignore
-        team_tool_calls = []  # type: ignore
-
-        # Track processed tool calls to avoid duplicates
-        processed_tool_calls = set()
-
-        with Live(console=console) as live_console:
-            status = Status("Thinking...", spinner="aesthetic", speed=0.4, refresh_per_second=10)
-            live_console.update(status)
-            response_timer = Timer()
-            response_timer.start()
-            # Flag which indicates if the panels should be rendered
-            render = False
-            # Panels to be rendered
-            panels = [status]
-            # First render the message panel if the message is not None
-            if input and show_message:
-                render = True
-                # Convert message to a panel
-                message_content = get_text_from_message(input)
-                message_panel = create_panel(
-                    content=Text(message_content, style="green"),
-                    title="Message",
-                    border_style="cyan",
-                )
-                panels.append(message_panel)
-            if render:
-                live_console.update(Group(*panels))
-
-            # Get response from the team
-            stream_resp = self.run(  # type: ignore
-                input=input,
-                audio=audio,
-                images=images,
-                videos=videos,
-                files=files,
-                stream=True,
-                stream_intermediate_steps=stream_intermediate_steps,
-                session_id=session_id,
-                session_state=session_state,
-                user_id=user_id,
-                knowledge_filters=knowledge_filters,
-                **kwargs,
-            )
-
-            team_markdown = None
-            member_markdown = {}
-
-            # Dict to track member response panels by member_id
-            member_response_panels = {}
-
-            run_id = None
-            for resp in stream_resp:
-                run_id = resp.run_id
-                if team_markdown is None:
-                    if markdown:
-                        team_markdown = True
-                    else:
-                        team_markdown = False
-
-                    if self.response_model is not None:
-                        team_markdown = False
-
-                if isinstance(resp, tuple(get_args(TeamRunOutputEvent))):
-                    if resp.event == TeamRunEvent.run_content:
-                        if isinstance(resp.content, str):
-                            _response_content += resp.content
-                        elif self.response_model is not None and isinstance(resp.content, BaseModel):
-                            try:
-                                _response_content = JSON(resp.content.model_dump_json(exclude_none=True), indent=2)  # type: ignore
-                            except Exception as e:
-                                log_warning(f"Failed to convert response to JSON: {e}")
-                        if hasattr(resp, "thinking") and resp.thinking is not None:
-                            _response_thinking += resp.thinking
-                    if (
-                        hasattr(resp, "metadata")
-                        and resp.metadata is not None
-                        and resp.metadata.reasoning_steps is not None
-                    ):
-                        reasoning_steps = resp.metadata.reasoning_steps
-
-                    # Collect team tool calls, avoiding duplicates
-                    if isinstance(resp, ToolCallCompletedEvent) and resp.tool:
-                        tool = resp.tool
-                        # Generate a unique ID for this tool call
-                        if tool.tool_call_id:
-                            tool_id = tool.tool_call_id
-                        else:
-                            tool_id = str(hash(str(tool)))
-                        if tool_id not in processed_tool_calls:
-                            processed_tool_calls.add(tool_id)
-                            team_tool_calls.append(tool)
-
-                # Collect member tool calls, avoiding duplicates
-                if hasattr(resp, "member_responses") and resp.member_responses:
-                    for member_response in resp.member_responses:
-                        member_id = None
-                        if isinstance(member_response, RunOutput) and member_response.id is not None:
-                            member_id = member_response.id
-                        elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                            member_id = member_response.id
-
-                        if member_id and hasattr(member_response, "tools") and member_response.tools:
-                            if member_id not in member_tool_calls:
-                                member_tool_calls[member_id] = []
-
-                            for tool in member_response.tools:
-                                # Generate a unique ID for this tool call
-                                if tool.tool_call_id:
-                                    tool_id = tool.tool_call_id
-                                else:
-                                    tool_id = str(hash(str(tool)))
-                                if tool_id not in processed_tool_calls:
-                                    processed_tool_calls.add(tool_id)
-                                    member_tool_calls[member_id].append(tool)
-
-                response_content_stream: Union[str, Markdown] = _response_content
-                # Escape special tags before markdown conversion
-                if team_markdown:
-                    escaped_content = escape_markdown_tags(_response_content, tags_to_include_in_markdown)
-                    response_content_stream = Markdown(escaped_content)
-
-                # Create new panels for each chunk
-                panels = []
-
-                if input and show_message:
-                    render = True
-                    # Convert message to a panel
-                    message_content = get_text_from_message(input)
-                    message_panel = create_panel(
-                        content=Text(message_content, style="green"),
-                        title="Message",
-                        border_style="cyan",
-                    )
-                    panels.append(message_panel)
-
-                if len(reasoning_steps) > 0 and show_reasoning:
-                    render = True
-                    # Create panels for reasoning steps
-                    for i, step in enumerate(reasoning_steps, 1):
-                        reasoning_panel = self._build_reasoning_step_panel(i, step, show_full_reasoning)
-                        panels.append(reasoning_panel)
-
-                if len(_response_thinking) > 0:
-                    render = True
-                    # Create panel for thinking
-                    thinking_panel = create_panel(
-                        content=Text(_response_thinking),
-                        title=f"Thinking ({response_timer.elapsed:.1f}s)",
-                        border_style="green",
-                    )
-                    panels.append(thinking_panel)
-                elif _response_content == "":
-                    # Keep showing status if no content yet
-                    panels.append(status)
-
-                # Process member responses and their tool calls
-                for member_response in resp.member_responses if hasattr(resp, "member_responses") else []:
-                    member_id = None
-                    member_name = "Team Member"
-                    if isinstance(member_response, RunOutput) and member_response.id is not None:
-                        member_id = member_response.id
-                        member_name = self._get_member_name(member_id)
-                    elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                        member_id = member_response.team_id
-
-                        member_name = self._get_member_name(member_id)
-
-                    # If we have tool calls for this member, display them
-                    if member_id in member_tool_calls and member_tool_calls[member_id]:
-                        formatted_calls = format_tool_calls(member_tool_calls[member_id])
-                        if formatted_calls:
-                            console_width = console.width if console else 80
-                            panel_width = console_width + 30
-
-                            lines = []
-                            for call in formatted_calls:
-                                wrapped_call = textwrap.fill(f"• {call}", width=panel_width, subsequent_indent="  ")
-                                lines.append(wrapped_call)
-
-                            tool_calls_text = "\n\n".join(lines)
-
-                            member_tool_calls_panel = create_panel(
-                                content=tool_calls_text,
-                                title=f"{member_name} Tool Calls",
-                                border_style="yellow",
-                            )
-                            panels.append(member_tool_calls_panel)
-
-                    # Process member response content
-                    if self.show_members_responses and member_id is not None:
-                        show_markdown = False
-                        if markdown:
-                            show_markdown = True
-
-                        member_response_content = self._parse_response_content(
-                            member_response,
-                            tags_to_include_in_markdown,
-                            show_markdown=show_markdown,
-                        )
-
-                        member_response_panel = create_panel(
-                            content=member_response_content,
-                            title=f"{member_name} Response",
-                            border_style="magenta",
-                        )
-
-                        panels.append(member_response_panel)
-
-                        # Store for reference
-                        if member_id is not None:
-                            member_response_panels[member_id] = member_response_panel
-
-                # Add team tool calls panel if available (before the team response)
-                if team_tool_calls:
-                    formatted_calls = format_tool_calls(team_tool_calls)
-                    if formatted_calls:
-                        console_width = console.width if console else 80
-                        panel_width = console_width + 30
-
-                        lines = []
-                        # Create a set to track already added calls by their string representation
-                        added_calls = set()
-                        for call in formatted_calls:
-                            if call not in added_calls:
-                                added_calls.add(call)
-                                # Wrap the call text to fit within the panel
-                                wrapped_call = textwrap.fill(f"• {call}", width=panel_width, subsequent_indent="  ")
-                                lines.append(wrapped_call)
-
-                        # Join with blank lines between items
-                        tool_calls_text = "\n\n".join(lines)
-
-                        team_tool_calls_panel = create_panel(
-                            content=tool_calls_text,
-                            title="Team Tool Calls",
-                            border_style="yellow",
-                        )
-                        panels.append(team_tool_calls_panel)
-
-                # Add the team response panel at the end
-                if response_content_stream:
-                    render = True
-                    # Create panel for response
-                    response_panel = create_panel(
-                        content=response_content_stream,
-                        title=f"Response ({response_timer.elapsed:.1f}s)",
-                        border_style="blue",
-                    )
-                    panels.append(response_panel)
-
-                if render or len(panels) > 0:
-                    live_console.update(Group(*panels))
-
-            response_timer.stop()
-            run_response = self.get_run_response(run_id=run_id)  # type: ignore
-
-            # Add citations
-            if hasattr(resp, "citations") and resp.citations is not None and resp.citations.urls is not None:
-                md_content = "\n".join(
-                    f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                    for i, citation in enumerate(resp.citations.urls)
-                    if citation.url  # Only include citations with valid URLs
-                )
-                if md_content:  # Only create panel if there are citations
-                    citations_panel = create_panel(
-                        content=Markdown(md_content),
-                        title="Citations",
-                        border_style="green",
-                    )
-                    panels.append(citations_panel)
-                    live_console.update(Group(*panels))
-
-            if self.memory_manager is not None:
-                if self.memory_manager.memories_updated:
-                    memory_panel = create_panel(
-                        content=Text("Memories updated"),
-                        title="Memories",
-                        border_style="green",
-                    )
-                    panels.append(memory_panel)
-                    live_console.update(Group(*panels))
-
-            if self.session_summary_manager is not None and self.session_summary_manager.summaries_updated:
-                summary_panel = create_panel(
-                    content=Text("Session summary updated"),
-                    title="Session Summary",
-                    border_style="green",
-                )
-                panels.append(summary_panel)
-                live_console.update(Group(*panels))
-                self.session_summary_manager.summaries_updated = False
-
-            # Final update to remove the "Thinking..." status
-            panels = [p for p in panels if not isinstance(p, Status)]
-
-            if markdown:
-                for member in self.members:
-                    if isinstance(member, Agent) and member.id is not None:
-                        member_markdown[member.id] = True
-                    if isinstance(member, Team) and member.id is not None:
-                        member_markdown[member.id] = True
-
-            for member in self.members:
-                if member.response_model is not None and isinstance(member, Agent) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-                if member.response_model is not None and isinstance(member, Team) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-
-            # Final panels assembly - we'll recreate the panels from scratch to ensure correct order
-            final_panels = []
-
-            # Start with the message
-            if input and show_message:
-                message_content = get_text_from_message(input)
-                message_panel = create_panel(
-                    content=Text(message_content, style="green"),
-                    title="Message",
-                    border_style="cyan",
-                )
-                final_panels.append(message_panel)
-
-            # Add reasoning steps
-            if reasoning_steps and show_reasoning:
-                for i, step in enumerate(reasoning_steps, 1):
-                    reasoning_panel = self._build_reasoning_step_panel(i, step, show_full_reasoning)
-                    final_panels.append(reasoning_panel)
-
-            # Add thinking panel if available
-            if _response_thinking:
-                thinking_panel = create_panel(
-                    content=Text(_response_thinking),
-                    title=f"Thinking ({response_timer.elapsed:.1f}s)",
-                    border_style="green",
-                )
-                final_panels.append(thinking_panel)
-
-            # Add member tool calls and responses in correct order
-            for i, member_response in enumerate(run_response.member_responses):  # type: ignore
-                member_id = None
-                if isinstance(member_response, RunOutput) and member_response.id is not None:
-                    member_id = member_response.id
-                elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                    member_id = member_response.id
-
-                if member_id:
-                    # First add tool calls if any
-                    if member_id in member_tool_calls and member_tool_calls[member_id]:
-                        formatted_calls = format_tool_calls(member_tool_calls[member_id])
-                        if formatted_calls:
-                            console_width = console.width if console else 80
-                            panel_width = console_width + 30
-
-                            lines = []
-                            for call in formatted_calls:
-                                wrapped_call = textwrap.fill(f"• {call}", width=panel_width, subsequent_indent="  ")
-                                lines.append(wrapped_call)
-
-                            tool_calls_text = "\n\n".join(lines)
-
-                            member_name = self._get_member_name(member_id)
-                            member_tool_calls_panel = create_panel(
-                                content=tool_calls_text,
-                                title=f"{member_name} Tool Calls",
-                                border_style="yellow",
-                            )
-                            final_panels.append(member_tool_calls_panel)
-
-                    # Add reasoning steps if any
-                    reasoning_steps = []
-                    if member_response.metadata is not None and member_response.metadata.reasoning_steps is not None:
-                        reasoning_steps = member_response.metadata.reasoning_steps
-                    if reasoning_steps and show_reasoning:
-                        for j, step in enumerate(reasoning_steps, 1):
-                            member_reasoning_panel = self._build_reasoning_step_panel(
-                                j, step, show_full_reasoning, color="magenta"
-                            )
-                            final_panels.append(member_reasoning_panel)
-
-                    # Then add response
-                    show_markdown = False
-                    if isinstance(member_response, RunOutput) and member_response.id is not None:
-                        show_markdown = member_markdown.get(member_response.id, False)
-                    elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                        show_markdown = member_markdown.get(member_response.id, False)
-
-                    member_response_content = self._parse_response_content(
-                        member_response,
-                        tags_to_include_in_markdown,
-                        show_markdown=show_markdown,
-                    )
-
-                    member_name = "Team Member"
-                    if isinstance(member_response, RunOutput) and member_response.id is not None:
-                        member_name = self._get_member_name(member_response.id)
-                    elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                        member_name = self._get_member_name(member_response.id)
-
-                    member_response_panel = create_panel(
-                        content=member_response_content,
-                        title=f"{member_name} Response",
-                        border_style="magenta",
-                    )
-                    final_panels.append(member_response_panel)
-
-                    # Add citations if any
-                    if member_response.citations is not None and member_response.citations.urls is not None:
-                        md_content = "\n".join(
-                            f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                            for i, citation in enumerate(member_response.citations.urls)
-                            if citation.url  # Only include citations with valid URLs
-                        )
-                        if md_content:  # Only create panel if there are citations
-                            citations_panel = create_panel(
-                                content=Markdown(md_content),
-                                title="Citations",
-                                border_style="magenta",
-                            )
-                            final_panels.append(citations_panel)
-
-            # Add team tool calls before team response
-            if team_tool_calls:
-                formatted_calls = format_tool_calls(team_tool_calls)
-                if formatted_calls:
-                    console_width = console.width if console else 80
-                    panel_width = console_width + 30
-
-                    lines = []
-                    # Create a set to track already added calls by their string representation
-                    added_calls = set()
-                    for call in formatted_calls:
-                        if call not in added_calls:
-                            added_calls.add(call)
-                            # Wrap the call text to fit within the panel
-                            wrapped_call = textwrap.fill(f"• {call}", width=panel_width, subsequent_indent="  ")
-                            lines.append(wrapped_call)
-
-                    tool_calls_text = "\n\n".join(lines)
-
-                    team_tool_calls_panel = create_panel(
-                        content=tool_calls_text,
-                        title="Team Tool Calls",
-                        border_style="yellow",
-                    )
-                    final_panels.append(team_tool_calls_panel)
-
-            # Add team response
-            if _response_content:
-                response_content_stream = _response_content
-                if team_markdown:
-                    escaped_content = escape_markdown_tags(_response_content, tags_to_include_in_markdown)
-                    response_content_stream = Markdown(escaped_content)
-
-                response_panel = create_panel(
-                    content=response_content_stream,
-                    title=f"Response ({response_timer.elapsed:.1f}s)",
-                    border_style="blue",
-                )
-                final_panels.append(response_panel)
-
-            # Add team citations
-            if hasattr(resp, "citations") and resp.citations is not None and resp.citations.urls is not None:
-                md_content = "\n".join(
-                    f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                    for i, citation in enumerate(resp.citations.urls)
-                    if citation.url  # Only include citations with valid URLs
-                )
-                if md_content:  # Only create panel if there are citations
-                    citations_panel = create_panel(
-                        content=Markdown(md_content),
-                        title="Citations",
-                        border_style="green",
-                    )
-                    final_panels.append(citations_panel)
-
-            # Final update with correctly ordered panels
-            live_console.update(Group(*final_panels))
 
     async def aprint_response(
         self,
@@ -3161,7 +2359,8 @@ class Team:
             self.markdown = markdown
 
         if stream:
-            await self._aprint_response_stream(
+            await aprint_response_stream(
+                team=self,
                 input=input,
                 console=console,
                 show_message=show_message,
@@ -3181,7 +2380,7 @@ class Team:
                 **kwargs,
             )
         else:
-            await self._aprint_response(
+            await aprint_response(
                 input=input,
                 console=console,
                 show_message=show_message,
@@ -3200,766 +2399,6 @@ class Team:
                 **kwargs,
             )
 
-    async def _aprint_response(
-        self,
-        input: Optional[Union[List, Dict, str, Message, BaseModel, List[Message]]] = None,
-        console: Optional[Any] = None,
-        show_message: bool = True,
-        show_reasoning: bool = True,
-        show_full_reasoning: bool = False,
-        tags_to_include_in_markdown: Optional[Set[str]] = None,
-        session_id: Optional[str] = None,
-        session_state: Optional[Dict[str, Any]] = None,
-        user_id: Optional[str] = None,
-        audio: Optional[Sequence[Audio]] = None,
-        images: Optional[Sequence[Image]] = None,
-        videos: Optional[Sequence[Video]] = None,
-        files: Optional[Sequence[File]] = None,
-        markdown: bool = False,
-        knowledge_filters: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> None:
-        import textwrap
-
-        from rich.console import Group
-        from rich.json import JSON
-        from rich.live import Live
-        from rich.markdown import Markdown
-        from rich.status import Status
-        from rich.text import Text
-
-        from agno.utils.response import format_tool_calls
-
-        if not tags_to_include_in_markdown:
-            tags_to_include_in_markdown = {"think", "thinking"}
-
-        with Live(console=console) as live_console:
-            status = Status("Thinking...", spinner="aesthetic", speed=0.4, refresh_per_second=10)
-            live_console.update(status)
-
-            response_timer = Timer()
-            response_timer.start()
-            # Panels to be rendered
-            panels = [status]
-            # First render the message panel if the message is not None
-            if input and show_message:
-                # Convert message to a panel
-                message_content = get_text_from_message(input)
-                message_panel = create_panel(
-                    content=Text(message_content, style="green"),
-                    title="Message",
-                    border_style="cyan",
-                )
-                panels.append(message_panel)
-                live_console.update(Group(*panels))
-
-            # Run the agent
-            run_response: TeamRunOutput = await self.arun(  # type: ignore
-                input=input,
-                images=images,
-                audio=audio,
-                videos=videos,
-                files=files,
-                stream=False,
-                session_id=session_id,
-                session_state=session_state,
-                user_id=user_id,
-                knowledge_filters=knowledge_filters,
-                **kwargs,
-            )
-            response_timer.stop()
-
-            team_markdown = False
-            member_markdown = {}
-            if markdown:
-                for member in self.members:
-                    if isinstance(member, Agent) and member.id is not None:
-                        member_markdown[member.id] = True
-                    if isinstance(member, Team) and member.id is not None:
-                        member_markdown[member.id] = True
-                team_markdown = True
-
-            if self.response_model is not None:
-                team_markdown = False
-
-            for member in self.members:
-                if member.response_model is not None and isinstance(member, Agent) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-                if member.response_model is not None and isinstance(member, Team) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-
-            # Handle reasoning
-            reasoning_steps = []
-            if (
-                isinstance(run_response, TeamRunOutput)
-                and run_response.metadata is not None
-                and run_response.metadata.reasoning_steps is not None
-            ):
-                reasoning_steps = run_response.metadata.reasoning_steps
-
-            if len(reasoning_steps) > 0 and show_reasoning:
-                # Create panels for reasoning steps
-                for i, step in enumerate(reasoning_steps, 1):
-                    reasoning_panel = self._build_reasoning_step_panel(i, step, show_full_reasoning)
-                    panels.append(reasoning_panel)
-                live_console.update(Group(*panels))
-
-            if isinstance(run_response, TeamRunOutput) and run_response.thinking is not None:
-                # Create panel for thinking
-                thinking_panel = create_panel(
-                    content=Text(run_response.thinking),
-                    title=f"Thinking ({response_timer.elapsed:.1f}s)",
-                    border_style="green",
-                )
-                panels.append(thinking_panel)
-                live_console.update(Group(*panels))
-
-            if isinstance(run_response, TeamRunOutput):
-                # Handle member responses
-                if self.show_members_responses:
-                    for member_response in run_response.member_responses:
-                        # Handle member reasoning
-                        reasoning_steps = []
-                        if (
-                            isinstance(member_response, RunOutput)
-                            and member_response.metadata is not None
-                            and member_response.metadata.reasoning_steps is not None
-                        ):
-                            reasoning_steps.extend(member_response.metadata.reasoning_steps)
-
-                        if len(reasoning_steps) > 0 and show_reasoning:
-                            # Create panels for reasoning steps
-                            for i, step in enumerate(reasoning_steps, 1):
-                                member_reasoning_panel = self._build_reasoning_step_panel(
-                                    i, step, show_full_reasoning, color="magenta"
-                                )
-                                panels.append(member_reasoning_panel)
-
-                        # Add tool calls panel for member if available
-                        if hasattr(member_response, "tools") and member_response.tools:
-                            member_name = None
-                            if isinstance(member_response, RunOutput) and member_response.id is not None:
-                                member_name = self._get_member_name(member_response.id)
-                            elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                                member_name = self._get_member_name(member_response.id)
-
-                            if member_name:
-                                # Format tool calls
-                                formatted_calls = format_tool_calls(member_response.tools)
-                                if formatted_calls:
-                                    console_width = console.width if console else 80
-                                    panel_width = console_width + 30
-
-                                    lines = []
-                                    for call in formatted_calls:
-                                        wrapped_call = textwrap.fill(
-                                            f"• {call}", width=panel_width, subsequent_indent="  "
-                                        )
-                                        lines.append(wrapped_call)
-
-                                    tool_calls_text = "\n\n".join(lines)
-
-                                    member_tool_calls_panel = create_panel(
-                                        content=tool_calls_text,
-                                        title=f"{member_name} Tool Calls",
-                                        border_style="yellow",
-                                    )
-                                    panels.append(member_tool_calls_panel)
-                                    live_console.update(Group(*panels))
-
-                        show_markdown = False
-                        if isinstance(member_response, RunOutput) and member_response.id is not None:
-                            show_markdown = member_markdown.get(member_response.id, False)
-                        elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                            show_markdown = member_markdown.get(member_response.id, False)
-
-                        member_response_content: Union[str, JSON, Markdown] = self._parse_response_content(
-                            member_response,
-                            tags_to_include_in_markdown,
-                            show_markdown=show_markdown,
-                        )
-
-                        # Create panel for member response
-                        if isinstance(member_response, RunOutput) and member_response.id is not None:
-                            member_response_panel = create_panel(
-                                content=member_response_content,
-                                title=f"{self._get_member_name(member_response.agent_id)} Response",
-                                border_style="magenta",
-                            )
-                        elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                            member_response_panel = create_panel(
-                                content=member_response_content,
-                                title=f"{self._get_member_name(member_response.team_id)} Response",
-                                border_style="magenta",
-                            )
-                        panels.append(member_response_panel)
-
-                        if member_response.citations is not None and member_response.citations.urls is not None:
-                            md_content = "\n".join(
-                                f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                                for i, citation in enumerate(member_response.citations.urls)
-                                if citation.url  # Only include citations with valid URLs
-                            )
-                            if md_content:
-                                citations_panel = create_panel(
-                                    content=Markdown(md_content),
-                                    title="Citations",
-                                    border_style="magenta",
-                                )
-                                panels.append(citations_panel)
-
-                    live_console.update(Group(*panels))
-
-                # Add team level tool calls panel if available
-                if run_response.tools:
-                    formatted_calls = format_tool_calls(run_response.tools)
-                    if formatted_calls:
-                        console_width = console.width if console else 80
-                        # Allow for panel borders and padding
-                        panel_width = console_width + 30
-
-                        lines = []
-                        for call in formatted_calls:
-                            # Wrap the call text to fit within the panel
-                            wrapped_call = textwrap.fill(f"• {call}", width=panel_width, subsequent_indent="  ")
-                            lines.append(wrapped_call)
-
-                        tool_calls_text = "\n\n".join(lines)
-
-                        team_tool_calls_panel = create_panel(
-                            content=tool_calls_text,
-                            title="Team Tool Calls",
-                            border_style="yellow",
-                        )
-                        panels.append(team_tool_calls_panel)
-                        live_console.update(Group(*panels))
-
-                response_content_batch: Union[str, JSON, Markdown] = self._parse_response_content(
-                    run_response, tags_to_include_in_markdown, show_markdown=team_markdown
-                )
-
-                # Create panel for response
-                response_panel = create_panel(
-                    content=response_content_batch,
-                    title=f"Response ({response_timer.elapsed:.1f}s)",
-                    border_style="blue",
-                )
-                panels.append(response_panel)
-
-                # Add citations
-                if run_response.citations is not None and run_response.citations.urls is not None:
-                    md_content = "\n".join(
-                        f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                        for i, citation in enumerate(run_response.citations.urls)
-                        if citation.url  # Only include citations with valid URLs
-                    )
-                    if md_content:  # Only create panel if there are citations
-                        citations_panel = create_panel(
-                            content=Markdown(md_content),
-                            title="Citations",
-                            border_style="green",
-                        )
-                        panels.append(citations_panel)
-
-                if self.memory_manager is not None:
-                    if self.memory_manager.memories_updated:
-                        memory_panel = create_panel(
-                            content=Text("Memories updated"),
-                            title="Memories",
-                            border_style="green",
-                        )
-                        panels.append(memory_panel)
-
-                if self.session_summary_manager is not None and self.session_summary_manager.summaries_updated:
-                    summary_panel = create_panel(
-                        content=Text("Session summary updated"),
-                        title="Session Summary",
-                        border_style="green",
-                    )
-                    panels.append(summary_panel)
-                    self.session_summary_manager.summaries_updated = False
-
-            # Final update to remove the "Thinking..." status
-            panels = [p for p in panels if not isinstance(p, Status)]
-            live_console.update(Group(*panels))
-
-    async def _aprint_response_stream(
-        self,
-        input: Optional[Union[List, Dict, str, Message, BaseModel, List[Message]]] = None,
-        console: Optional[Any] = None,
-        show_message: bool = True,
-        show_reasoning: bool = True,
-        show_full_reasoning: bool = False,
-        tags_to_include_in_markdown: Optional[Set[str]] = None,
-        session_id: Optional[str] = None,
-        session_state: Optional[Dict[str, Any]] = None,
-        user_id: Optional[str] = None,
-        audio: Optional[Sequence[Audio]] = None,
-        images: Optional[Sequence[Image]] = None,
-        videos: Optional[Sequence[Video]] = None,
-        files: Optional[Sequence[File]] = None,
-        markdown: bool = False,
-        stream_intermediate_steps: bool = False,  # type: ignore
-        knowledge_filters: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> None:
-        import textwrap
-
-        from rich.console import Group
-        from rich.json import JSON
-        from rich.live import Live
-        from rich.markdown import Markdown
-        from rich.status import Status
-        from rich.text import Text
-
-        if not tags_to_include_in_markdown:
-            tags_to_include_in_markdown = {"think", "thinking"}
-
-        stream_intermediate_steps = True  # With streaming print response, we need to stream intermediate steps
-
-        _response_content: str = ""
-        _response_thinking: str = ""
-        reasoning_steps: List[ReasoningStep] = []
-
-        # Track tool calls by member and team
-        member_tool_calls = {}  # type: ignore
-        team_tool_calls: List[ToolExecution] = []
-
-        # Track processed tool calls to avoid duplicates
-        processed_tool_calls = set()
-
-        # Initialize final_panels here
-        final_panels = []  # type: ignore
-
-        with Live(console=console) as live_console:
-            status = Status("Thinking...", spinner="aesthetic", speed=0.4, refresh_per_second=10)
-            live_console.update(status)
-            response_timer = Timer()
-            response_timer.start()
-            # Flag which indicates if the panels should be rendered
-            render = False
-            # Panels to be rendered
-            panels = [status]
-            # First render the message panel if the message is not None
-            if input and show_message:
-                render = True
-                # Convert message to a panel
-                message_content = get_text_from_message(input)
-                message_panel = create_panel(
-                    content=Text(message_content, style="green"),
-                    title="Message",
-                    border_style="cyan",
-                )
-                panels.append(message_panel)
-            if render:
-                live_console.update(Group(*panels))
-
-            # Get response from the team
-            team_markdown = None
-            member_markdown = {}
-
-            run_id = None
-            async for resp in self.arun(  # type: ignore
-                input=input,
-                audio=audio,
-                images=images,
-                videos=videos,
-                files=files,
-                stream=True,
-                stream_intermediate_steps=stream_intermediate_steps,
-                session_id=session_id,
-                session_state=session_state,
-                user_id=user_id,
-                knowledge_filters=knowledge_filters,
-                **kwargs,
-            ):
-                run_id = resp.run_id
-                if team_markdown is None:
-                    if markdown:
-                        team_markdown = True
-                    else:
-                        team_markdown = False
-
-                    if self.response_model is not None:
-                        team_markdown = False
-
-                if isinstance(resp, tuple(get_args(TeamRunOutputEvent))):
-                    if resp.event == TeamRunEvent.run_content:
-                        if isinstance(resp.content, str):
-                            _response_content += resp.content
-                        elif self.response_model is not None and isinstance(resp.content, BaseModel):
-                            try:
-                                _response_content = JSON(resp.content.model_dump_json(exclude_none=True), indent=2)  # type: ignore
-                            except Exception as e:
-                                log_warning(f"Failed to convert response to JSON: {e}")
-                        if resp.thinking is not None:
-                            _response_thinking += resp.thinking
-                    if (
-                        hasattr(resp, "metadata")
-                        and resp.metadata is not None
-                        and resp.metadata.reasoning_steps is not None
-                    ):
-                        reasoning_steps = resp.metadata.reasoning_steps
-
-                    # Collect team tool calls, avoiding duplicates
-                    if isinstance(resp, ToolCallCompletedEvent) and resp.tool:
-                        tool = resp.tool
-                        # Generate a unique ID for this tool call
-                        if tool.tool_call_id is not None:
-                            tool_id = tool.tool_call_id
-                        else:
-                            tool_id = str(hash(str(tool)))
-                        if tool_id not in processed_tool_calls:
-                            processed_tool_calls.add(tool_id)
-                            team_tool_calls.append(tool)
-
-                # Collect member tool calls, avoiding duplicates
-                if hasattr(resp, "member_responses") and resp.member_responses:
-                    for member_response in resp.member_responses:
-                        member_id = None
-                        if isinstance(member_response, RunOutput) and member_response.id is not None:
-                            member_id = member_response.id
-                        elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                            member_id = member_response.id
-
-                        if member_id and hasattr(member_response, "tools") and member_response.tools:
-                            if member_id not in member_tool_calls:
-                                member_tool_calls[member_id] = []
-
-                            for tool in member_response.tools:
-                                if tool.tool_call_id is not None:
-                                    tool_id = tool.tool_call_id
-                                else:
-                                    tool_id = str(hash(str(tool)))
-                                if tool_id not in processed_tool_calls:
-                                    processed_tool_calls.add(tool_id)
-                                    member_tool_calls[member_id].append(tool)
-
-                response_content_stream: Union[str, Markdown] = _response_content
-                # Escape special tags before markdown conversion
-                if team_markdown:
-                    escaped_content = escape_markdown_tags(_response_content, tags_to_include_in_markdown)
-                    response_content_stream = Markdown(escaped_content)
-
-                # Create new panels for each chunk
-                panels = [status]
-
-                if input and show_message:
-                    render = True
-                    # Convert message to a panel
-                    message_content = get_text_from_message(input)
-                    message_panel = create_panel(
-                        content=Text(message_content, style="green"),
-                        title="Message",
-                        border_style="cyan",
-                    )
-                    panels.append(message_panel)
-                if render:
-                    live_console.update(Group(*panels))
-
-                if len(reasoning_steps) > 0 and show_reasoning:
-                    render = True
-                    # Create panels for reasoning steps
-                    for i, step in enumerate(reasoning_steps, 1):
-                        reasoning_panel = self._build_reasoning_step_panel(i, step, show_full_reasoning)
-                        panels.append(reasoning_panel)
-                if render:
-                    live_console.update(Group(*panels))
-
-                if len(_response_thinking) > 0:
-                    render = True
-                    # Create panel for thinking
-                    thinking_panel = create_panel(
-                        content=Text(_response_thinking),
-                        title=f"Thinking ({response_timer.elapsed:.1f}s)",
-                        border_style="green",
-                    )
-                    panels.append(thinking_panel)
-                if render:
-                    live_console.update(Group(*panels))
-
-                # Add tool calls panel if available
-                if (
-                    resp is not None
-                    and hasattr(resp, "tool")
-                    and resp.tool is not None
-                    and resp.event == RunEvent.tool_call_started
-                ):
-                    render = True
-                    # Create bullet points for each tool call
-                    tool_calls_content = Text()
-                    formatted_tool_call = format_tool_calls([resp.tool])
-                    tool_calls_content.append(f"• {formatted_tool_call}\n")
-
-                    tool_calls_panel = create_panel(
-                        content=tool_calls_content.plain.rstrip(),
-                        title="Tool Calls",
-                        border_style="yellow",
-                    )
-                    panels.append(tool_calls_panel)
-
-                if response_content_stream:
-                    render = True
-                    # Create panel for response
-                    response_panel = create_panel(
-                        content=response_content_stream,
-                        title=f"Response ({response_timer.elapsed:.1f}s)",
-                        border_style="blue",
-                    )
-                    panels.append(response_panel)
-                if render:
-                    live_console.update(Group(*panels))
-            response_timer.stop()
-
-            run_response = self.get_run_response(run_id=run_id)  # type: ignore
-
-            # Add citations
-            if hasattr(resp, "citations") and resp.citations is not None and resp.citations.urls is not None:
-                md_content = "\n".join(
-                    f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                    for i, citation in enumerate(resp.citations.urls)
-                    if citation.url  # Only include citations with valid URLs
-                )
-                if md_content:  # Only create panel if there are citations
-                    citations_panel = create_panel(
-                        content=Markdown(md_content),
-                        title="Citations",
-                        border_style="green",
-                    )
-                    panels.append(citations_panel)
-                    live_console.update(Group(*panels))
-
-            if self.memory_manager is not None:
-                if self.memory_manager.memories_updated:
-                    memory_panel = create_panel(
-                        content=Text("Memories updated"),
-                        title="Memories",
-                        border_style="green",
-                    )
-                    panels.append(memory_panel)
-                    live_console.update(Group(*panels))
-
-            if self.session_summary_manager is not None and self.session_summary_manager.summaries_updated:
-                summary_panel = create_panel(
-                    content=Text("Session summary updated"),
-                    title="Session Summary",
-                    border_style="green",
-                )
-                panels.append(summary_panel)
-                live_console.update(Group(*panels))
-                self.session_summary_manager.summaries_updated = False
-
-            # Final update to remove the "Thinking..." status
-            panels = [p for p in panels if not isinstance(p, Status)]
-
-            if markdown:
-                for member in self.members:
-                    if isinstance(member, Agent) and member.id is not None:
-                        member_markdown[member.id] = True  # type: ignore
-                    if isinstance(member, Team) and member.id is not None:
-                        member_markdown[member.id] = True  # type: ignore
-
-            for member in self.members:
-                if member.response_model is not None and isinstance(member, Agent) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-                if member.response_model is not None and isinstance(member, Team) and member.id is not None:
-                    member_markdown[member.id] = False  # type: ignore
-
-            # Final panels assembly - we'll recreate the panels from scratch to ensure correct order
-            final_panels = []
-
-            # Start with the message
-            if input and show_message:
-                message_content = get_text_from_message(input)
-                message_panel = create_panel(
-                    content=Text(message_content, style="green"),
-                    title="Message",
-                    border_style="cyan",
-                )
-                final_panels.append(message_panel)
-
-            # Add reasoning steps
-            if reasoning_steps and show_reasoning:
-                for i, step in enumerate(reasoning_steps, 1):
-                    reasoning_panel = self._build_reasoning_step_panel(i, step, show_full_reasoning)
-                    final_panels.append(reasoning_panel)
-
-            # Add thinking panel if available
-            if _response_thinking:
-                thinking_panel = create_panel(
-                    content=Text(_response_thinking),
-                    title=f"Thinking ({response_timer.elapsed:.1f}s)",
-                    border_style="green",
-                )
-                final_panels.append(thinking_panel)
-
-            # Add member tool calls and responses in correct order
-            for i, member_response in enumerate(run_response.member_responses):  # type: ignore
-                member_id = None
-                if isinstance(member_response, RunOutput) and member_response.id is not None:
-                    member_id = member_response.id
-                elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                    member_id = member_response.id
-
-                if member_id:
-                    # First add tool calls if any
-                    if member_id in member_tool_calls and member_tool_calls[member_id]:
-                        formatted_calls = format_tool_calls(member_tool_calls[member_id])
-                        if formatted_calls:
-                            console_width = console.width if console else 80
-                            panel_width = console_width + 30
-
-                            lines = []
-                            # Create a set to track already added calls by their string representation
-                            added_calls = set()
-                            for call in formatted_calls:
-                                if call not in added_calls:
-                                    added_calls.add(call)
-                                    # Wrap the call text to fit within the panel
-                                    wrapped_call = textwrap.fill(f"• {call}", width=panel_width, subsequent_indent="  ")
-                                    lines.append(wrapped_call)
-
-                            tool_calls_text = "\n\n".join(lines)
-
-                            member_name = self._get_member_name(member_id)
-                            member_tool_calls_panel = create_panel(
-                                content=tool_calls_text,
-                                title=f"{member_name} Tool Calls",
-                                border_style="yellow",
-                            )
-                            final_panels.append(member_tool_calls_panel)
-
-                    # Add reasoning steps if any
-                    reasoning_steps = []
-                    if member_response.metadata is not None and member_response.metadata.reasoning_steps is not None:
-                        reasoning_steps = member_response.metadata.reasoning_steps
-                    if reasoning_steps and show_reasoning:
-                        for j, step in enumerate(reasoning_steps, 1):
-                            member_reasoning_panel = self._build_reasoning_step_panel(
-                                j, step, show_full_reasoning, color="magenta"
-                            )
-                            final_panels.append(member_reasoning_panel)
-
-                    # Then add response
-                    show_markdown = False
-                    if isinstance(member_response, RunOutput) and member_response.id is not None:
-                        show_markdown = member_markdown.get(member_response.id, False)
-                    elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                        show_markdown = member_markdown.get(member_response.id, False)
-
-                    member_response_content = self._parse_response_content(
-                        member_response,
-                        tags_to_include_in_markdown,
-                        show_markdown=show_markdown,
-                    )
-
-                    member_name = "Team Member"
-                    if isinstance(member_response, RunOutput) and member_response.id is not None:
-                        member_name = self._get_member_name(member_response.id)
-                    elif isinstance(member_response, TeamRunOutput) and member_response.id is not None:
-                        member_name = self._get_member_name(member_response.id)
-
-                    member_response_panel = create_panel(
-                        content=member_response_content,
-                        title=f"{member_name} Response",
-                        border_style="magenta",
-                    )
-                    final_panels.append(member_response_panel)
-
-                    # Add citations if any
-                    if member_response.citations is not None and member_response.citations.urls is not None:
-                        md_content = "\n".join(
-                            f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                            for i, citation in enumerate(member_response.citations.urls)
-                            if citation.url  # Only include citations with valid URLs
-                        )
-                        if md_content:  # Only create panel if there are citations
-                            citations_panel = create_panel(
-                                content=Markdown(md_content),
-                                title="Citations",
-                                border_style="magenta",
-                            )
-                            final_panels.append(citations_panel)
-
-            # Add team tool calls before team response
-            if team_tool_calls:
-                formatted_calls = format_tool_calls(team_tool_calls)
-                if formatted_calls:
-                    console_width = console.width if console else 80
-                    panel_width = console_width + 30
-
-                    lines = []
-                    # Create a set to track already added calls by their string representation
-                    added_calls = set()
-                    for call in formatted_calls:
-                        if call not in added_calls:
-                            added_calls.add(call)
-                            # Wrap the call text to fit within the panel
-                            wrapped_call = textwrap.fill(f"• {call}", width=panel_width, subsequent_indent="  ")
-                            lines.append(wrapped_call)
-
-                    tool_calls_text = "\n\n".join(lines)
-
-                    team_tool_calls_panel = create_panel(
-                        content=tool_calls_text,
-                        title="Team Tool Calls",
-                        border_style="yellow",
-                    )
-                    final_panels.append(team_tool_calls_panel)
-
-            # Add team response
-            if _response_content:
-                response_content_stream = _response_content
-                if team_markdown:
-                    escaped_content = escape_markdown_tags(_response_content, tags_to_include_in_markdown)
-                    response_content_stream = Markdown(escaped_content)
-
-                response_panel = create_panel(
-                    content=response_content_stream,
-                    title=f"Response ({response_timer.elapsed:.1f}s)",
-                    border_style="blue",
-                )
-                final_panels.append(response_panel)
-
-            # Add team citations
-            if hasattr(resp, "citations") and resp.citations is not None and resp.citations.urls is not None:
-                md_content = "\n".join(
-                    f"{i + 1}. [{citation.title or citation.url}]({citation.url})"
-                    for i, citation in enumerate(resp.citations.urls)
-                    if citation.url  # Only include citations with valid URLs
-                )
-                if md_content:  # Only create panel if there are citations
-                    citations_panel = create_panel(
-                        content=Markdown(md_content),
-                        title="Citations",
-                        border_style="green",
-                    )
-                    final_panels.append(citations_panel)
-
-            # Final update with correctly ordered panels
-            live_console.update(Group(*final_panels))
-
-    def _build_reasoning_step_panel(
-        self, step_idx: int, step: ReasoningStep, show_full_reasoning: bool = False, color: str = "green"
-    ):
-        from rich.text import Text
-
-        # Build step content
-        step_content = Text.assemble()
-        if step.title is not None:
-            step_content.append(f"{step.title}\n", "bold")
-        if step.action is not None:
-            step_content.append(Text.from_markup(f"[bold]Action:[/bold] {step.action}\n", style="dim"))
-        if step.result is not None:
-            step_content.append(Text.from_markup(step.result, style="dim"))
-
-        if show_full_reasoning:
-            # Add detailed reasoning information if available
-            if step.reasoning is not None:
-                step_content.append(Text.from_markup(f"\n[bold]Reasoning:[/bold] {step.reasoning}", style="dim"))
-            if step.confidence is not None:
-                step_content.append(Text.from_markup(f"\n[bold]Confidence:[/bold] {step.confidence}", style="dim"))
-        return create_panel(content=step_content, title=f"Reasoning step {step_idx}", border_style=color)
 
     def _get_member_name(self, entity_id: str) -> str:
         for member in self.members:
@@ -6637,7 +5076,7 @@ class Team:
                     knowledge_filters=knowledge_filters
                     if not member_agent.knowledge_filters and member_agent.knowledge
                     else None,
-                    refresh_session_before_write=True,
+                    yield_run_response=True,
                 )
                 member_agent_run_response = None
                 async for member_agent_run_response_event in member_agent_run_response_stream:
